Restructure Node_Hardware MCP for better organization and dependency management
- Remove node_hardware subfolder from /src and move files directly under /src
- Update pyproject.toml to match HDF5 structure with clean dependencies
- Fix all import paths throughout the codebase to work with flat structure
- Remove unused dependencies (fastapi, uvicorn, pydantic, requests, mcp[cli])
- Keep only essential dependencies: fastmcp, python-dotenv, psutil
- Add hardware-focused keywords from README content
- Update project structure to be consistent with other MCPs
