+++ mcps/Node_Hardware/pyproject.toml
@@ -1,45 +1,31 @@
+[project]
+name = "node-hardware-mcp"
+version = "1.0.0"
+description = "Node Hardware MCP - Comprehensive Hardware Monitoring and System Analysis for LLMs with real-time performance metrics"
+readme = "README.md"
+requires-python = ">=3.10"
+license = "MIT"
+authors = [
+    { name = "IoWarp Scientific MCPs", email = "contact@iowarp.org" }
+]
+
+dependencies = [
+  "fastmcp",
+  "python-dotenv>=1.0.0",
+  "psutil>=5.9.0"
+]
+
+keywords = ["hardware-monitoring", "system-analysis", "performance-metrics", "node-information", "ssh-monitoring", "remote-hardware", "mcp", "llm-integration", "infrastructure-monitoring", "distributed-systems"]
+
+[tool.uv]
+dev-dependencies = [
+    "pytest>=8.4.0",
+    "pytest-asyncio>=0.21.0"
+]
+
+[project.scripts]
+node-hardware-mcp = "server:main"
+
 [build-system]
 requires = ["setuptools>=64.0", "wheel"]
 build-backend = "setuptools.build_meta"
-
-[project]
-name = "node-hardware-mcp"
-version = "0.1.0"
-description = "Node Hardware MCP server implementation using Model Context Protocol"
-readme = "README.md"
-requires-python = ">=3.10.12"
-dependencies = [
-  "fastmcp>=0.1.0",
-  "mcp[cli]>=0.1.0",
-  "pytest-asyncio>=1.0.0",
-  "python-dotenv>=1.0.0",
-  "psutil>=5.9.0",
-  "fastapi>=0.95.0",
-  "uvicorn>=0.21.0",
-  "pydantic>=1.10.0",
-  "pytest>=7.2.0",
-  "requests>=2.28.0"
-]
-
-[project.optional-dependencies]
-test = ["pytest>=7.0.0", "httpx"]
-
-[project.scripts]
-node-hardware-mcp = "node_hardware.server:main"
-
-[tool.setuptools.packages.find]
-where = ["src"]
-include = ["node_hardware*"]
-
-[tool.setuptools.package-dir]
-"node_hardware" = "src/node_hardware"
-
-[dependency-groups]
-dev = [
-    "pytest>=8.4.0",
-]
-
-[tool.uv.workspace]
-members = [
-    "node-hardware-mcp",
-]
