+++ mcps/Node_Hardware/src/capabilities/__init__.py
