+++ mcps/Node_Hardware/src/capabilities/memory_info.py
@@ -3,7 +3,7 @@ Memory information capabilities.
 Handles memory reporting and detailed memory information.
 """
 import psutil
-from .utils import format_bytes, format_percentage
+from capabilities.utils import format_bytes, format_percentage
 
 
 def get_memory_info() -> dict:
