+++ mcps/Node_Hardware/src/capabilities/utils.py
