--- mcps/Node_Hardware/src/node_hardware/__init__.py
@@ -1,3 +1,0 @@
-"""
-Node Hardware MCP Server package.
-"""
