+++ mcps/Node_Hardware/src/utils/output_formatter.py
